refactor(headers): move per-output parse into ShieldedOutput.deserialize
ShieldedOutput owned serialize()/serializeSighash() but not the decode —
the per-output parse was inlined in ShieldedOutputsHeader.deserialize, so
the model could write itself but not read itself, and the wire layout was
described in two files that had to change in lockstep.

Add static ShieldedOutput.deserialize(buf): [ShieldedOutput, Buffer] next
to serialize, and have the header loop over it. This mirrors hathor-core,
where serialize_shielded_output and deserialize_shielded_output both live
in the output module and the header just iterates. The per-output
hathor-core size caps (range proof / script / surjection) move with the
parse — they're per-output field validations, so they belong on the model;
the header keeps only the num_outputs bounds.

Single source of truth for the output wire format, and a reusable
single-output codec for later PRs (e.g. reconstructing from raw bytes).

Adds direct model deserialize tests (round-trip AmountShielded/FullShielded,
buffer chaining/leftover, unsupported mode, range-proof size cap). The
header tests still exercise the same path through delegation. 148
header+model tests pass; lint and tsc clean.
