+++ src/models/shielded_output.ts
@@ -5,9 +5,14 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import { intToBytes } from '../utils/buffer';
+import { intToBytes, unpackToInt } from '../utils/buffer';
 import { ShieldedOutputMode } from '../shielded/types';
 import { OutputValueType } from '../types';
+import {
+  MAX_RANGE_PROOF_SIZE,
+  MAX_SURJECTION_PROOF_SIZE,
+  MAX_SHIELDED_OUTPUT_SCRIPT_SIZE,
+} from '../constants';
 
 const EPHEMERAL_PUBKEY_SIZE = 33;
 
@@ -132,6 +137,110 @@ class ShieldedOutput {
     arr.push(this.getValidatedEphemeralPubkey());
 
     return arr;
+  }
+
+  /**
+   * Parse a single shielded output from the wire (inverse of `serialize`).
+   *
+   * Co-located with `serialize` so the wire layout has a single source of
+   * truth (mirrors hathor-core, where `serialize_shielded_output` and
+   * `deserialize_shielded_output` both live in the output module and the
+   * header just loops over them). Returns the parsed output plus the
+   * remaining buffer so callers (e.g. `ShieldedOutputsHeader.deserialize`)
+   * can chain across a list.
+   *
+   * The on-chain form carries no plaintext value — it's recovered later via
+   * rewind — so the model is built with `value = 0n`.
+   */
+  static deserialize(srcBuf: Buffer): [ShieldedOutput, Buffer] {
+    let buf = Buffer.from(srcBuf);
+
+    // Mode (1 byte)
+    if (buf.length < 1) throw new Error('Truncated shielded output: missing mode byte');
+    let mode: number;
+    // eslint-disable-next-line prefer-const
+    [mode, buf] = unpackToInt(1, false, buf);
+
+    // Commitment (33 bytes)
+    if (buf.length < 33) throw new Error('Truncated shielded output: missing commitment');
+    const commitment = Buffer.from(buf.subarray(0, 33));
+    buf = buf.subarray(33);
+
+    // Range proof (2 bytes length + variable)
+    if (buf.length < 2) throw new Error('Truncated shielded output: missing range proof length');
+    let rpLen: number;
+    // eslint-disable-next-line prefer-const
+    [rpLen, buf] = unpackToInt(2, false, buf);
+    // hathor-core validation: range proof size cannot exceed MAX_RANGE_PROOF_SIZE.
+    if (rpLen > MAX_RANGE_PROOF_SIZE)
+      throw new Error(`range proof size ${rpLen} exceeds maximum ${MAX_RANGE_PROOF_SIZE}`);
+    if (buf.length < rpLen) throw new Error('Truncated shielded output: incomplete range proof');
+    const rangeProof = Buffer.from(buf.subarray(0, rpLen));
+    buf = buf.subarray(rpLen);
+
+    // Script (2 bytes length + variable)
+    if (buf.length < 2) throw new Error('Truncated shielded output: missing script length');
+    let scriptLen: number;
+    // eslint-disable-next-line prefer-const
+    [scriptLen, buf] = unpackToInt(2, false, buf);
+    // hathor-core validation: shielded output script size cannot exceed MAX_SHIELDED_OUTPUT_SCRIPT_SIZE.
+    if (scriptLen > MAX_SHIELDED_OUTPUT_SCRIPT_SIZE)
+      throw new Error(
+        `script size ${scriptLen} exceeds maximum ${MAX_SHIELDED_OUTPUT_SCRIPT_SIZE}`
+      );
+    if (buf.length < scriptLen) throw new Error('Truncated shielded output: incomplete script');
+    const script = Buffer.from(buf.subarray(0, scriptLen));
+    buf = buf.subarray(scriptLen);
+
+    let tokenData = 0;
+    let assetCommitment: Buffer | undefined;
+    let surjectionProof: Buffer | undefined;
+
+    if (mode === ShieldedOutputMode.AMOUNT_SHIELDED) {
+      // Token data (1 byte)
+      if (buf.length < 1) throw new Error('Truncated AmountShielded output: missing token_data');
+      [tokenData, buf] = unpackToInt(1, false, buf);
+    } else if (mode === ShieldedOutputMode.FULLY_SHIELDED) {
+      // Asset commitment (33 bytes)
+      if (buf.length < 33)
+        throw new Error('Truncated FullShielded output: missing asset commitment');
+      assetCommitment = Buffer.from(buf.subarray(0, 33));
+      buf = buf.subarray(33);
+
+      // Surjection proof (2 bytes length + variable)
+      if (buf.length < 2)
+        throw new Error('Truncated FullShielded output: missing surjection proof length');
+      let spLen: number;
+      [spLen, buf] = unpackToInt(2, false, buf);
+      // hathor-core validation: surjection proof size cannot exceed MAX_SURJECTION_PROOF_SIZE.
+      if (spLen > MAX_SURJECTION_PROOF_SIZE)
+        throw new Error(
+          `surjection proof size ${spLen} exceeds maximum ${MAX_SURJECTION_PROOF_SIZE}`
+        );
+      if (buf.length < spLen)
+        throw new Error('Truncated FullShielded output: incomplete surjection proof');
+      surjectionProof = Buffer.from(buf.subarray(0, spLen));
+      buf = buf.subarray(spLen);
+    } else {
+      throw new Error(`Unsupported shielded output mode: ${mode}`);
+    }
+
+    // Ephemeral pubkey (33 bytes)
+    if (buf.length < 33) throw new Error('Truncated shielded output: missing ephemeral pubkey');
+    const ephemeralPubkey = Buffer.from(buf.subarray(0, 33));
+    buf = buf.subarray(33);
+
+    const output = new ShieldedOutput(
+      mode,
+      commitment,
+      rangeProof,
+      tokenData,
+      script,
+      ephemeralPubkey,
+      0n, // value is not stored on-chain; it's recovered via rewind
+      { assetCommitment, surjectionProof }
+    );
+    return [output, buf];
   }
 
   /**
